requirements.txt
playwright
playwright-stealth
google-generativeai
groq

pydantic
tqdm
python-dotenv
requests
beautifulsoup4
